# api/docs/swagger.yaml
basePath: /api/v1/
definitions:
  domain.Admin:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      password:
        type: string
      schoolID:
        type: string
    type: object
  domain.ContactInfo:
    properties:
      address:
        type: string
      businessName:
        type: string
      email:
        type: string
      phone:
        type: string
      registrationNumber:
        type: string
    type: object
  domain.Course:
    properties:
      code:
        type: string
      color:
        type: string
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      imageUrl:
        type: string
      name:
        type: string
      published:
        type: boolean
      updatedAt:
        type: string
    type: object
...
  domain.Fondy:
    properties:
      connected:
        type: boolean
      merchantId:
        type: string
      merchantPassword:
        type: string
    type: object
...
  domain.Pages:
    properties:
      confidential:
        type: string
      newsletterConsent:
        type: string
      serviceAgreement:
        type: string
    type: object
...
  domain.School:
    properties:
      admins:
        items:
          $ref: '#/definitions/domain.Admin'
        type: array
      courses:
        items:
          $ref: '#/definitions/domain.Course'
        type: array
      description:
        type: string
      id:
        type: string
      name:
        type: string
      registeredAt:
        type: string
      settings:
        $ref: '#/definitions/domain.Settings'
      subtitle:
        type: string
    type: object
  domain.SendPulse:
    properties:
      connected:
        type: boolean
      id:
        type: string
      listId:
        type: string
      secret:
        type: string
    type: object
...
  domain.Settings:
    properties:
      color:
        type: string
      contactInfo:
        $ref: '#/definitions/domain.ContactInfo'
      disableRegistration:
        type: boolean
      domains:
        items:
          type: string
        type: array
      fondy:
        $ref: '#/definitions/domain.Fondy'
      googleAnalyticsCode:
        type: string
      logo:
        type: string
      pages:
        $ref: '#/definitions/domain.Pages'
      sendpulse:
        $ref: '#/definitions/domain.SendPulse'
      showPaymentImages:
        type: boolean
    type: object
...
  v1.createSchoolInput:
    properties:
      name:
        type: string
    required:
    - name
    type: object
...
  v1.refreshInput:
    properties:
      token:
        type: string
    required:
    - token
    type: object
  v1.response:
    properties:
      message:
        type: string
    type: object
...
  v1.signInInput:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  v1.tokenResponse:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
    type: object
...
  v1.userSignUpInput:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        type: string
    required:
    - email
    - name
    - password
    - phone
    type: object
host: localhost:8000
info:
  contact: {}
  description: REST API for UltraThreads App
  title: UltraThreads API
  version: "1.0"
paths:
...
  /users/auth/refresh:
    post:
      consumes:
      - application/json
      description: user refresh tokens
      parameters:
      - description: sign up info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.refreshInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.tokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      summary: User Refresh Tokens
      tags:
      - users-auth
  /users/schools:
    post:
      consumes:
      - application/json
      description: user create school
      parameters:
      - description: school info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.createSchoolInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.School'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - UsersAuth: []
      summary: User Create New School
      tags:
      - users-schools
  /users/sign-in:
    post:
      consumes:
      - application/json
      description: user sign in
      parameters:
      - description: sign up info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.signInInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.tokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      summary: User SignIn
      tags:
      - users-auth
  /users/sign-up:
    post:
      consumes:
      - application/json
      description: create user account
      parameters:
      - description: sign up info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.userSignUpInput'
      produces:
      - application/json
      responses:
        "201":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      summary: User SignUp
      tags:
      - users-auth
  /users/verify/{code}:
    post:
      consumes:
      - application/json
      description: user verify registration
      parameters:
      - description: verification code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.tokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - UsersAuth: []
      summary: User Verify Registration
      tags:
      - users-auth
securityDefinitions:
  AdminAuth:
    in: header
    name: Authorization
    type: apiKey
  StudentsAuth:
    in: header
    name: Authorization
    type: apiKey
  UsersAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"